修改了活动部分代码 Signed-off-by: ypfengWei <642514910@qq.com>

+++ src/main/java/com/shanduo/party/service/impl/ActivityServiceImpl.java
@@ -103,18 +103,25 @@ public class ActivityServiceImpl implements ActivityService {
 	}
 	
 	@Override
-	public List<Map<String, Object>> selectByActivityId(String activityId, Integer pageNum, Integer pageSize, Integer userId) {
+	public Map<String, Object> selectByActivityId(String activityId, Integer pageNum, Integer pageSize, Integer userId) {
 		int totalrecord = shanduoActivityMapper.selectByScoreActivityCount(activityId);
 		Page page = new Page(totalrecord, pageSize, pageNum);
 		pageNum = (page.getPageNum() - 1) * page.getPageSize();
-		List<Map<String, Object>> resultMap = shanduoActivityMapper.selectByActivityId(activityId, pageNum, page.getPageSize());
-		for (Map<String, Object> map : resultMap) {
+		List<Map<String, Object>> resultList = shanduoActivityMapper.selectByActivityId(activityId, pageNum, page.getPageSize());
+		int joinActivity = 0;
+		for (Map<String, Object> map : resultList) {
 			if(map.get("id").equals(userId)) {
-				map.put("joinActivity", 1);//已报名
-			} else {
-				map.put("joinActivity", 0);//未报名
-			}
+				joinActivity = 1;
+				break;
+			} 
 		}
+		for (Map<String, Object> map : resultList) {
+			map.put("joinActivity", joinActivity);
+		}
+		Map<String, Object> resultMap = new HashMap<>(3);
+		resultMap.put("page", page.getPageNum());
+		resultMap.put("totalPage", page.getTotalPage());
+		resultMap.put("list", resultList);
 		return resultMap;
 	}
 	
@@ -184,6 +191,11 @@ public class ActivityServiceImpl implements ActivityService {
 			log.error("删除需求失败");
 			throw new RuntimeException();
 		}
+		int score = activityScoreMapper.deleteByActivityId(activityId);
+		if(score < 1) {
+			log.error("删除参加用户信息失败");
+			throw new RuntimeException();
+		}
 		return 1;
 	}
 	
@@ -219,7 +231,7 @@ public class ActivityServiceImpl implements ActivityService {
 	public Map<String, Object> selectByNearbyUserId(String lon,String lat, Integer pageNum, Integer pageSize) {
 		long longtime = System.currentTimeMillis();
 		Date time = new Date(longtime - 1000*60*60*12);
-		shanduoActivityMapper.updateById(time); // 取消活动置顶
+		shanduoActivityMapper.updateById(time); // 置顶超过12小时的活动取消置顶
 		Double[] doubles = LocationUtils.getDoubles(lon, lat);
 		int totalrecord = shanduoActivityMapper.selectByNearbyUserIdCount(doubles[0], doubles[1], doubles[2], doubles[3]);
 		Page page = new Page(totalrecord, pageSize, pageNum);
@@ -378,12 +390,18 @@ public class ActivityServiceImpl implements ActivityService {
 	
 	@Override
 	public int deleteByUserId(String activityId, Integer userId) {
-		int i = activityScoreMapper.deleteByUserId(activityId, userId);
-		if(i < 1) {
-			log.error("取消活动失败");
+		ShanduoActivity activity = shanduoActivityMapper.selectByPrimaryKey(activityId);
+		if(activity.getActivityStartTime().getTime() < System.currentTimeMillis()) {
+			log.error("该活动已过期");
 			throw new RuntimeException();
+		} else {
+			int i = activityScoreMapper.deleteByUserId(activityId, userId);
+			if(i < 1) {
+				log.error("取消活动失败");
+				throw new RuntimeException();
+			}
 		}
-		return 0;
+		return 1;
 	}
 	
 	public List<ActivityInfo> activity(List<ActivityInfo> resultList, String lon, String lat){
@@ -441,9 +459,12 @@ public class ActivityServiceImpl implements ActivityService {
     				activityInfo.setWomanNumber("0/"+activityInfo.getWomanNumber());
     			}
     		}
+    		
     		SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd HH:mm");  
             String startString = activityInfo.getActivityStartTime();  
             String cutoffString = activityInfo.getActivityCutoffTime(); 
+            Long newstartTime = get(startString, null, 2);
+            activityInfo.setNewStartTime(newstartTime);
             String startTime = startString.substring(0,16).replace("-", "/").replace(" ", "/");
             String cutoffTime = cutoffString.substring(0,16).replace("-", "/").replace(" ", "/");
             Long time = get(null, "yyyy-MM-dd 00:00", 1);
